refactor(l5): move the agent-context-pack sections, budget, guidance, repo-facts and pointers-candidates modules from src/ into server/tools/legacy-context-pack (issue 864)
Five agent-context-pack sibling modules move out of src/ into a new
server/tools/legacy-context-pack/ directory. They are the members of the
cluster that read no process.env and do not import agent-context-pack.ts, so
they move without dragging the import cycle formed by agent-context-pack.ts,
-durable-lane.ts and -durable-memory.ts, which stay in src/ for a later lane.

The target directory is deliberately separate from server/tools/, which
already holds the hardened context-pack-*.ts rewrite plus its own
agent-context-pack.ts. Nothing is merged with that rewrite.

Moved (git mv, old -> new):
- src/tools/agent-context-pack-sections.ts ->
  server/tools/legacy-context-pack/agent-context-pack-sections.ts
- src/tools/agent-context-pack-budget.ts ->
  server/tools/legacy-context-pack/agent-context-pack-budget.ts
- src/tools/agent-context-pack-guidance.ts ->
  server/tools/legacy-context-pack/agent-context-pack-guidance.ts
- src/tools/agent-context-pack-repo-facts.ts ->
  server/tools/legacy-context-pack/agent-context-pack-repo-facts.ts
- src/tools/agent-context-pack-pointers-candidates.ts ->
  server/tools/legacy-context-pack/agent-context-pack-pointers-candidates.ts

Import repointing inside the moved files:
- ./agent-context-pack-sections.ts stays a sibling path in the new directory.
- ../observability/index.ts -> ../../observability/log-surface.ts (the server/
  surface that re-exports logger and describeError).
- ./repo-facts.ts -> ../../domain/repo-facts.ts (the src/ path was itself an L5
  shim over that module).
- ./search-brain.ts and ./agent-context-pack-durable-memory.ts keep their src/
  paths, rewritten relative: search-brain is moving in a parallel lane and
  durable-memory has not moved.

Seams the server/ lint standard required, mechanical and behavior-preserving:
- agent-context-pack-budget.ts: the checkpoint binary search inside
  fitDurableLaneSection is hoisted into a named fitCheckpointText helper,
  dropping the function's complexity under the rule.
- agent-context-pack-guidance.ts: the dead lifecycleOverflow variable is
  removed, and the promoted-row selection and item assembly inside
  loadGuidanceSection are hoisted into selectPromotedRows and
  collectGuidanceItems, which also removes both over-deep blocks.
- agent-context-pack-repo-facts.ts: the per-row admission logic is hoisted into
  repoFactRow with an explicit skipped/omitted/kept result, collected by
  collectRepoFactItems, and the no-active-repo empty state into
  noActiveRepoFragment.

Each old src/ path is now an L5 shim re-exporting the moved module, so the
remaining src/ importers (agent-context-pack.ts, src/tools/index.ts) and every
test keep their import lines untouched. No server/ or scripts/ importer of
these five modules exists, and the survey for text readers found only Python
doc-comment references, not path guards.

The seven agent-context-pack test files stay at their src/ paths per the M1
exception and test the moved code through the shims, unmodified: 101 pass /
0 fail before the change at origin/main and 101 pass / 0 fail after.

Runtime closure of src/ under server/main.ts: 17 -> 12.

issue 864
